Profitable liquidation of repaid positions
Previously, the liquidation of repaid positions was not profitable for the system as all the proceeds got paid out to the position owner. Now, we are ensuring that the bid amount is shared between the system and the position owner. This closes a major attack vector which allowed a position owner to avoid losses by repaying a challenged position.