Introduce global minting limit
Note: with this change, it will become possible for fast cloners of a new position to use the limit before the owner of the original position, making the minting process for new positions more competitive. The correct way for the initial owner to address this is to set the interest so high that the demand for loans with that collateral stays below the limit.